% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_rows.R
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapse redundant probes into gene symbols}
\usage{
collapse_rows(expr, probe_col, gene_col, data_table = FALSE,
  method = c("max_mean", "min_mean", "col_mean", "col_median"))
}
\arguments{
\item{expr}{Object of class \code{data.frame} containing expression values,
a column with probe symbols and a column with gene symbols}

\item{probe_col}{Character string specifying the name of the column containing 
probe names. Can also take rownames as input using the string "rownames".}

\item{gene_col}{Character string specifying the name of the column containing 
gene symbols to be used to collapse.}

\item{data_table}{Logical. If \code{TRUE} will return a \code{data.table} object.
Defaults to \code{FALSE}.}

\item{method}{Method used to collapse the probes. Can take one of "max_mean", 
"min_mean", "col_mean" or "col_median".}
}
\value{
Object of class \code{data.frame} or \code{data.table}
}
\description{
This function collapses redundant probe expression rows into single rows
based on repeated gene symbols in a given column.
}
\examples{
# Create mock probe expression data
mock <- CEMiTool::expr
mock$Symbol <- rownames(mock)
rownames(mock) <- NULL
mock <- mock[1:(nrow(mock)/2),]
mock2 <- mock[,names(mock) != "Symbol"]
mock2 <- mock2 * 2
mock2$Symbol <- mock$Symbol
mock <- rbind(mock, mock2)
mock <- mock[order(mock$Symbol), ]
rownames(mock) <- paste0("probe_", rownames(mock))

# Collapse redundant probes
expr_col <- collapse_rows(mock, probe_col="rownames", gene_col="Symbol", method="max_mean")

}
