% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a CSBL project directory with the standard structure}
\usage{
create_project(directory, force = FALSE)
}
\arguments{
\item{directory}{The path into which the folders should be created.}

\item{force}{If directory exists, processing will not be stopped.}
}
\value{
An appropriate file environment for CSBL projects
}
\description{
This function will create a directory with the standard folders for 
a project in the CSBL: data, intermediate, results and src.
}
\examples{
# Create project file structure
create_project(directory <- "~/test", force=FALSE)

}
