% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ev_filter.R
\name{ev_filter}
\alias{ev_filter}
\title{Expression and variance filtering}
\usage{
ev_filter(expr, genenum, prop)
}
\arguments{
\item{expr}{Object of class \code{data.frame} containing gene expression data.
genes must be in the rows, samples in the columns, and any identifiers must be 
in the rownames.}

\item{genenum}{Number of genes to keep in the resulting object.}

\item{prop}{Proportion of most variant genes to keep in the resulting object.}
}
\value{
Object of class \code{data.frame} with filtered expression values
}
\description{
This function allows you to filter a gene expression dataset
by the number of genes you want in the resulting dataset.
}
\examples{
# Get example expression data
expr <- CEMiTool::expr
# Select 60\% most expressed and 2000 most variant genes
expr_f <- ev_filter(expr, 2000, 0.6)

}
