% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_multiple_names.R
\name{search_multiple_names}
\alias{search_multiple_names}
\title{search_multiple_names}
\usage{
search_multiple_names(vector_of_interest, query_vector, perl = TRUE,
  return_index = FALSE)
}
\arguments{
\item{vector_of_interest}{The vector that you want to perform a search in.}

\item{query_vector}{A vector of the words that you want to match (at least one of them)}

\item{perl}{Logical (TRUE or FALSE). Indicates whether perl-compatible regex will be used.
Defaults to \code{FALSE}.}

\item{return_index}{Logical (TRUE or FALSE). Indicates whether the search will return indexes or the values themselves in the vector.
Defaults to false (thus values are returned).}
}
\description{
Wrapper to search match in a R vector for many words
}
\examples{
# Vector de interesse para realizar a busca
csblers <- c("Andre Nicolau", "Andre Guilherme", "Deney Muleque Doido", "Thiago Hirata", "Tiago Lubiana")
# Vector contendo os valores que serão buscados
query <- c("Thiago", "Tiago")
# Comando grep para buscar os valores de interesse
search_multiple_names(vector_of_interest, query_vector = query)
search_multiple_names(vector_of_interest, query_vector = query, return_index = TRUE)
}
